+++ bot/common/command/utilities.py
@@ -1,0 +1,2 @@
+def humanize(number: int) -> str:
+    return f"{number:,}"