+++ bot/module/economics/plugin/command/collect.py
@@ -2,10 +2,9 @@ import random
 
 import crescent
 
-from bot.common.command import commands, cooldowns, embeds
-from bot.common.command.utility import utilities
+from bot.common.command import commands, cooldowns, embeds, utilities
 from bot.common.plugin import plugins
-from bot.module.economics.shop import shops
+from bot.common.shop import shops
 
 plugin = plugins.Plugin()
 
@@ -87,26 +86,3 @@ class Command(commands.Command):
         )
 
         await context.respond(embed=embed)
-
-
-# MIT License
-#
-# Copyright (c) 2023 elaresai
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.