+++ bot/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added API
 
+### Removed
+
+- Removed code comments with license
+
 ### Fixed
 
 - Fixed a real number required error